cloud: move common uri glob helpers to cloud
Release note: none.
